Refactor route organization to resource/views pattern
Reorganize 17 route files from flat structure to scalable resource/views pattern:

## Changes Made

### File Structure
- **Resources**: API routes moved to `resources/api/admin/`
- **Views**: UI routes organized by feature in `views/`
  - `views/auth/` - Authentication pages (login, signup, confirm, etc.)
  - `views/admin/` - Admin interface (users, setup, test-coverage)
  - `views/account/` - User account pages
  - `views/public/` - Public pages (home, logout, protected)

### Route Configuration
- Updated `app/routes.ts` imports while preserving all existing URLs
- Fixed TypeScript import paths in moved files
- Maintained existing route hierarchy and layout nesting

### Validation
- All TypeScript compilation passes (`npm run tsc`)
- All 147 unit tests pass (`npm run test`)
- No functional changes or regressions

This refactor improves maintainability and provides a foundation for scaling to 100+ routes while keeping the codebase organized and developer-friendly.
