+++ app/routes/resources/api/admin/users.test.tsx
@@ -20,7 +20,7 @@ vi.mock('loglevel', () => ({
 
 // Import after mocks are set up
 import '~/__tests__/mocks/admin'
-import { loader, action } from './api.admin.users'
+import { loader, action } from './users'
 import { createClient } from '~/lib/supabase/client'
 
 describe('API Admin Users Route', () => {
