+++ app/routes/views/public/logout.tsx
