+++ app/routes/views/public/protected.tsx
